backend/requirements.txt
fastapi==0.118.3
starlette==0.48.0
uvicorn==0.30.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
motor==3.3.2
pydantic==2.5.2
pydantic-settings==2.1.0
openai==1.99.9
google-generativeai==0.8.5
litellm==1.77.7
aiohttp==3.13.0
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
Faker==20.1.0
redis==5.0.1
email-validator==2.1.0
reportlab==4.0.7
pandas==2.1.4
openpyxl==3.1.2
emergentintegrations
